requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
langchain-core>=0.2.0
langchain-google-genai>=0.1.0
langchain>=0.2.0
todoist-api-python>=2.2.0
pydantic>=2.0.0
